서버를 dll라이브러리로 수정, 서버앱, 테스트클라이언트 프로젝트 추가

+++ GameServer2/struct.h
@@ -37,7 +37,7 @@ struct IOCompletionData
     void SetBuffer(const std::vector<uint8_t>& _sendBuffer)
     {
         wsabuf.buf = buffer;
-        wsabuf.len = _sendBuffer.size();
+        wsabuf.len = static_cast<ULONG>(_sendBuffer.size());
         
         // 데이터 범위 체크 최대 255크기 까지
         assert(wsabuf.len < IOBUFFER_SIZE);
